store: split out refresh/action support to store_action.go
